+++ fe/fe-core/src/main/java/org/apache/doris/datasource/iceberg/source/IcebergScanNode.java
@@ -23,6 +23,7 @@ import org.apache.doris.analysis.TableSnapshot;
 import org.apache.doris.analysis.TupleDescriptor;
 import org.apache.doris.catalog.Env;
 import org.apache.doris.catalog.TableIf;
+import org.apache.doris.common.Config;
 import org.apache.doris.common.DdlException;
 import org.apache.doris.common.UserException;
 import org.apache.doris.common.security.authentication.ExecutionAuthenticator;
@@ -58,34 +59,20 @@ import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.Lists;
 import org.apache.commons.lang3.exception.ExceptionUtils;
-import org.apache.doris.common.Config;
-import org.apache.iceberg.BaseFileScanTask;
 import org.apache.iceberg.BaseTable;
-import org.apache.iceberg.DataFile;
 import org.apache.iceberg.DeleteFile;
-import com.google.common.collect.Sets;
 import org.apache.iceberg.FileContent;
 import org.apache.iceberg.FileScanTask;
 import org.apache.iceberg.ManifestFile;
-import org.apache.iceberg.ManifestFiles;
-import org.apache.iceberg.ManifestReader;
 import org.apache.iceberg.MetadataColumns;
 import org.apache.iceberg.PartitionData;
 import org.apache.iceberg.PartitionSpec;
-import org.apache.iceberg.SchemaParser;
 import org.apache.iceberg.Snapshot;
-import org.apache.iceberg.StructLike;
 import org.apache.iceberg.Table;
 import org.apache.iceberg.TableScan;
-import org.apache.iceberg.expressions.Evaluator;
 import org.apache.iceberg.expressions.Expression;
-import org.apache.iceberg.expressions.Expressions;
-import org.apache.iceberg.expressions.InclusiveMetricsEvaluator;
-import org.apache.iceberg.expressions.Projections;
-import org.apache.iceberg.expressions.ResidualEvaluator;
 import org.apache.iceberg.io.CloseableIterable;
 import org.apache.iceberg.io.CloseableIterator;
-import org.apache.iceberg.metrics.ScanMetrics;
 import org.apache.iceberg.types.Conversions;
 import org.apache.iceberg.util.TableScanUtil;
 import org.apache.logging.log4j.LogManager;
@@ -98,7 +85,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.OptionalLong;
-import java.util.Set;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.atomic.AtomicReference;
 
@@ -130,6 +116,8 @@ public class IcebergScanNode extends FileQueryScanNode {
     // get them in doInitialize() to ensure internal consistency of ScanNode
     private Map<StorageProperties.Type, StorageProperties> storagePropertiesMap;
     private Map<String, String> backendStorageProperties;
+    // Planner for Iceberg file scan tasks with optional manifest cache support
+    private final IcebergFileScanPlanner fileScanPlanner = new IcebergFileScanPlanner();
 
     // for test
     @VisibleForTesting
@@ -375,48 +363,29 @@ public class IcebergScanNode extends FileQueryScanNode {
 
     private CloseableIterable<FileScanTask> planFileScanTask(TableScan scan) {
         long targetSplitSize = getRealFileSplitSize(0);
-        
+
         if (!Config.enable_iceberg_manifest_cache) {
             // Cache disabled, use original logic
             return TableScanUtil.splitFiles(scan.planFiles(), targetSplitSize);
         }
-        
-        // Use cache logic
-        return planFileScanTaskWithCache(scan, targetSplitSize);
-    }
 
-    private CloseableIterable<FileScanTask> planFileScanTaskWithCache(TableScan scan, long targetSplitSize) {
+        // Use cache logic
         try {
             IcebergManifestCache manifestCache = Env.getCurrentEnv()
                     .getExtMetaCacheMgr()
                     .getIcebergMetadataCache()
                     .getManifestCache();
-            
+
             Snapshot snapshot = scan.snapshot();
             if (snapshot == null) {
                 return TableScanUtil.splitFiles(scan.planFiles(), targetSplitSize);
             }
-            
-            // Get table name for cache tracking
-            String tableName = icebergTable.name();
-            if (tableName == null || tableName.isEmpty()) {
-                ExternalTable table = (ExternalTable) desc.getTable();
-                tableName = table.getCatalog().getName() + "." + table.getDbName() + "." + table.getName();
-            }
-            
-            // Get data manifests and delete manifests
-            List<ManifestFile> dataManifests = Lists.newArrayList(
-                    IcebergUtils.getMatchingManifest(
-                            snapshot.dataManifests(icebergTable.io()),
-                            icebergTable.specs(),
-                            scan.filter()));
-            List<ManifestFile> deleteManifests = Lists.newArrayList(
-                    snapshot.deleteManifests(icebergTable.io()));
-            
-            // Process data manifests with cache
-            CloseableIterable<FileScanTask> tasks = planDataManifestsWithCache(
-                    dataManifests, deleteManifests, manifestCache, scan, tableName);
 
+            String tableName = resolveTableName();
+            IcebergFileScanPlanner.PlanContext ctx = new IcebergFileScanPlanner.PlanContext(
+                    scan, icebergTable, snapshot, manifestCache, tableName);
+
+            CloseableIterable<FileScanTask> tasks = fileScanPlanner.planWithCache(ctx);
             return TableScanUtil.splitFiles(tasks, targetSplitSize);
         } catch (Exception e) {
             LOG.warn("Failed to use manifest cache, fallback to normal plan", e);
@@ -424,401 +393,13 @@ public class IcebergScanNode extends FileQueryScanNode {
         }
     }
 
-    /**
-     * Process data manifests with cache support
-     */
-    private CloseableIterable<FileScanTask> planDataManifestsWithCache(
-            List<ManifestFile> dataManifests,
-            List<ManifestFile> deleteManifests,
-            IcebergManifestCache manifestCache,
-            TableScan scan,
-            String tableName) {
-
-        List<FileScanTask> cachedTasks = new ArrayList<>();
-        List<ManifestFile> uncachedManifests = new ArrayList<>();
-
-        // Separate cached and uncached manifests
-        for (ManifestFile manifestFile : dataManifests) {
-            Set<DataFile> dataFiles = manifestCache.getDataFiles(manifestFile.path());
-            if (dataFiles != null && !dataFiles.isEmpty()) {
-                // Cache hit: create FileScanTask from cached DataFiles
-                List<FileScanTask> tasks = createFileScanTasksFromCachedDataFiles(
-                        manifestFile, dataFiles, deleteManifests, scan);
-                cachedTasks.addAll(tasks);
-            } else {
-                // Cache miss: mark for reading
-                manifestCache.prepareCache(manifestFile.path(), tableName);
-                uncachedManifests.add(manifestFile);
-            }
+    private String resolveTableName() {
+        String tableName = icebergTable.name();
+        if (tableName == null || tableName.isEmpty()) {
+            ExternalTable table = (ExternalTable) desc.getTable();
+            tableName = table.getCatalog().getName() + "." + table.getDbName() + "." + table.getName();
         }
-
-        // Process uncached manifests: read and fill cache
-        CloseableIterable<FileScanTask> uncachedTasks = null;
-        if (!uncachedManifests.isEmpty()) {
-            uncachedTasks = planFileScanTaskFromManifests(
-                    uncachedManifests, deleteManifests, manifestCache, scan, tableName);
-        }
-
-        // Merge cached and uncached results
-        return mergeFileScanTasks(cachedTasks, uncachedTasks);
-    }
-
-    /**
-     * Create FileScanTask from cached DataFiles
-     */
-    private List<FileScanTask> createFileScanTasksFromCachedDataFiles(
-            ManifestFile manifestFile,
-            Set<DataFile> cachedDataFiles,
-            List<ManifestFile> deleteManifests,
-            TableScan scan) {
-
-        List<FileScanTask> tasks = new ArrayList<>();
-
-        // Get partition and file filters
-        Expression dataFilter = scan.filter();
-        PartitionSpec spec = icebergTable.specs().get(manifestFile.partitionSpecId());
-
-        // Create evaluators for filtering
-        // 1. Partition filter: project dataFilter to partition expression
-        // Reference: ManifestGroup.java line 252-257
-        Expression partitionFilter = Projections.inclusive(spec, scan.isCaseSensitive())
-                .project(dataFilter);
-        Evaluator partitionEvaluator = new Evaluator(
-                spec.partitionType(), partitionFilter, scan.isCaseSensitive());
-
-        // 2. File filter: use InclusiveMetricsEvaluator for file-level metrics
-        // filtering
-        // Reference: ManifestReader.java line 232, 248-250
-        InclusiveMetricsEvaluator fileEvaluator = new InclusiveMetricsEvaluator(
-                icebergTable.schema(), dataFilter, scan.isCaseSensitive());
-
-        // Get cached delete files for all delete manifests
-        Set<DeleteFile> cachedDeleteFiles = Sets.newHashSet();
-        for (ManifestFile deleteManifest : deleteManifests) {
-            Set<DeleteFile> deleteFiles = Env.getCurrentEnv()
-                    .getExtMetaCacheMgr()
-                    .getIcebergMetadataCache()
-                    .getManifestCache()
-                    .getDeleteFiles(deleteManifest.path());
-            if (deleteFiles != null) {
-                cachedDeleteFiles.addAll(deleteFiles);
-            }
-        }
-
-        // Create TaskContext
-        // Note: For cached files, we don't ignore residuals to maintain correctness
-        // Reference: DataTableScan.java line 82-84
-        boolean ignoreResiduals = false;
-        // Calculate dropStats: if there are equality deletes, we need to keep stats
-        // Reference: ManifestGroup.java line 187-190
-        boolean dropStats = false; // Simplified: always keep stats for cached files
-        Set<Integer> columnsToKeepStats = null; // Simplified: keep all stats
-        TaskContextHelper taskContext = createTaskContext(
-                spec, scan.filter(), scan.isCaseSensitive(), ignoreResiduals, dropStats, columnsToKeepStats);
-
-        for (DataFile dataFile : cachedDataFiles) {
-            // 1. Apply partition filter
-            // Reference: ManifestReader.java line 248:
-            // evaluator.eval(entry.file().partition())
-            if (!partitionEvaluator.eval(dataFile.partition())) {
-                continue;
-            }
-
-            // 2. Apply file filter (based on metrics)
-            // Reference: ManifestReader.java line 249: metricsEvaluator.eval(entry.file())
-            if (!fileEvaluator.eval(dataFile)) {
-                continue;
-            }
-
-            // 3. Find matching DeleteFiles
-            DeleteFile[] deleteFilesArray = findDeleteFilesForDataFile(
-                    dataFile, cachedDeleteFiles);
-
-            // 4. Create FileScanTask
-            FileScanTask task = createFileScanTaskFromDataFile(
-                    dataFile, deleteFilesArray, taskContext);
-            tasks.add(task);
-        }
-
-        return tasks;
-    }
-
-    /**
-     * Create FileScanTask from DataFile
-     */
-    private FileScanTask createFileScanTaskFromDataFile(
-            DataFile dataFile,
-            DeleteFile[] deleteFiles,
-            TaskContextHelper taskContext) {
-
-        // Copy DataFile (defensive copy)
-        // Reference: ManifestGroup.createFileScanTasks() line 365
-        // Note: ContentFileUtil.copy() is package-private, so we use direct copy
-        // methods
-        // For simplicity, we always copy with stats (can be optimized later)
-        DataFile copiedDataFile = taskContext.shouldKeepStats()
-                ? dataFile.copy()
-                : dataFile.copyWithoutStats();
-
-        // Create BaseFileScanTask (using public constructor)
-        return new BaseFileScanTask(
-                copiedDataFile,
-                deleteFiles != null ? deleteFiles : new DeleteFile[0],
-                taskContext.schemaAsString(),
-                taskContext.specAsString(),
-                taskContext.residuals());
-    }
-
-    /**
-     * Create TaskContext for creating FileScanTask
-     * Reference: ManifestGroup.TaskContext (line 379-399)
-     */
-    private TaskContextHelper createTaskContext(
-            PartitionSpec spec,
-            Expression dataFilter,
-            boolean caseSensitive,
-            boolean ignoreResiduals,
-            boolean dropStats,
-            Set<Integer> columnsToKeepStats) {
-
-        // Create ResidualEvaluator
-        // Reference: ManifestGroup.java line 176-183
-        Expression filter = ignoreResiduals ? Expressions.alwaysTrue() : dataFilter;
-        ResidualEvaluator residuals = ResidualEvaluator.of(spec, filter, caseSensitive);
-
-        // Serialize Schema and Spec
-        // Reference: ManifestGroup.TaskContext constructor (line 395-396)
-        String schemaAsString = SchemaParser.toJson(spec.schema());
-        String specAsString = org.apache.iceberg.PartitionSpecParser.toJson(spec);
-
-        return new TaskContextHelper(
-                schemaAsString,
-                specAsString,
-                residuals,
-                dropStats,
-                        columnsToKeepStats,
-                ScanMetrics.noop());
-    }
-
-    /**
-     * Find DeleteFiles for a DataFile
-     */
-    private DeleteFile[] findDeleteFilesForDataFile(
-            DataFile dataFile,
-            Set<DeleteFile> cachedDeleteFiles) {
-
-        if (cachedDeleteFiles == null || cachedDeleteFiles.isEmpty()) {
-            return new DeleteFile[0];
-        }
-
-        List<DeleteFile> matchingDeletes = new ArrayList<>();
-        long dataFileSequence = dataFile.dataSequenceNumber();
-
-        for (DeleteFile deleteFile : cachedDeleteFiles) {
-            // Check sequence number
-            if (deleteFile.dataSequenceNumber() < dataFileSequence) {
-                continue;
-            }
-
-            // Match based on DeleteFile type
-            if (deleteFile.content() == FileContent.EQUALITY_DELETES) {
-                // Equality deletes: match partition
-                if (matchesPartition(dataFile, deleteFile)) {
-                    matchingDeletes.add(deleteFile);
-                }
-            } else if (deleteFile.content() == FileContent.POSITION_DELETES) {
-                // Position deletes: match file path or partition
-                if (matchesFileOrPartition(dataFile, deleteFile)) {
-                    matchingDeletes.add(deleteFile);
-                }
-            }
-        }
-
-        return matchingDeletes.toArray(new DeleteFile[0]);
-    }
-
-    /**
-     * Check if DeleteFile matches DataFile's partition
-     */
-    private boolean matchesPartition(DataFile dataFile, DeleteFile deleteFile) {
-        StructLike dataPartition = dataFile.partition();
-        StructLike deletePartition = deleteFile.partition();
-
-        if (dataPartition == null || deletePartition == null) {
-            return false;
-        }
-
-        return dataPartition.equals(deletePartition);
-    }
-
-    /**
-     * Check if DeleteFile matches DataFile's file path or partition
-     */
-    private boolean matchesFileOrPartition(DataFile dataFile, DeleteFile deleteFile) {
-        // For position deletes, check if file path matches
-        // Note: ContentFileUtil.referencedDataFile() is package-private, so we use a
-        // simplified check
-        // Check if delete file's path bounds match data file path
-        if (deleteFile.lowerBounds() != null && deleteFile.upperBounds() != null) {
-            // Simplified: if bounds exist and are equal, might reference a specific file
-            // In practice, we'd need to check the actual referenced file path
-        }
-
-        // Otherwise check partition
-        return matchesPartition(dataFile, deleteFile);
-    }
-
-    /**
-     * Process uncached manifests: read and fill cache
-     */
-    private CloseableIterable<FileScanTask> planFileScanTaskFromManifests(
-            List<ManifestFile> uncachedManifests,
-            List<ManifestFile> deleteManifests,
-            IcebergManifestCache manifestCache,
-            TableScan scan,
-            String tableName) {
-
-        // For uncached manifests, use TableScan.planFiles() and fill cache
-        // Note: We still use scan.planFiles() for uncached manifests to leverage
-        // Iceberg's filtering and planning logic
-        CloseableIterable<FileScanTask> tasks = scan.planFiles();
-
-        // Wrap to fill cache during iteration
-        return wrapTasksWithCacheFill(tasks, uncachedManifests, deleteManifests,
-                manifestCache, tableName);
-    }
-
-    /**
-     * Merge cached and uncached FileScanTasks
-     */
-    private CloseableIterable<FileScanTask> mergeFileScanTasks(
-            List<FileScanTask> cachedTasks,
-            CloseableIterable<FileScanTask> uncachedTasks) {
-
-        if (cachedTasks.isEmpty() && uncachedTasks == null) {
-            return CloseableIterable.empty();
-        }
-
-        if (cachedTasks.isEmpty()) {
-            return uncachedTasks;
-        }
-
-        if (uncachedTasks == null) {
-            return CloseableIterable.withNoopClose(cachedTasks);
-        }
-
-        // Merge both
-        return CloseableIterable.combine(
-                CloseableIterable.withNoopClose(cachedTasks),
-                uncachedTasks);
-    }
-
-    /**
-     * Wrap FileScanTask to collect DataFile and DeleteFile for caching
-     * For uncached manifests, read them using ManifestFiles.read() and fill cache
-     */
-    private CloseableIterable<FileScanTask> wrapTasksWithCacheFill(
-            CloseableIterable<FileScanTask> tasks,
-            List<ManifestFile> uncachedManifests,
-                    List<ManifestFile> deleteManifests,
-            IcebergManifestCache manifestCache,
-            String tableName) {
-        
-        // Read uncached manifests and fill cache
-        Map<String, Set<DataFile>> manifestDataFiles = new HashMap<>();
-        Map<String, Set<DeleteFile>> manifestDeleteFiles = new HashMap<>();
-        
-        // Read data manifests and fill cache
-        // Note: We use ManifestReader's iterator() which returns ContentFile directly
-        for (ManifestFile manifest : uncachedManifests) {
-            try {
-                ManifestReader<DataFile> reader = ManifestFiles.read(manifest, icebergTable.io(),
-                        icebergTable.specs());
-                Set<DataFile> dataFiles = Sets.newHashSet();
-                try (CloseableIterable<DataFile> files = reader) {
-                    for (DataFile file : files) {
-                        dataFiles.add(file);
-                    }
-                }
-                if (!dataFiles.isEmpty()) {
-                    manifestDataFiles.put(manifest.path(), dataFiles);
-                    manifestCache.putDataFiles(manifest.path(), dataFiles);
-                }
-            } catch (Exception e) {
-                LOG.warn("Failed to read manifest {} for cache", manifest.path(), e);
-            }
-        }
-
-        // Read delete manifests and fill cache
-        for (ManifestFile manifest : deleteManifests) {
-            try {
-                ManifestReader<DeleteFile> reader = ManifestFiles.readDeleteManifest(
-                        manifest, icebergTable.io(), icebergTable.specs());
-                Set<DeleteFile> deleteFiles = Sets.newHashSet();
-                try (CloseableIterable<DeleteFile> files = reader) {
-                    for (DeleteFile file : files) {
-                        deleteFiles.add(file);
-                    }
-                }
-                if (!deleteFiles.isEmpty()) {
-                    manifestDeleteFiles.put(manifest.path(), deleteFiles);
-                    manifestCache.putDeleteFiles(manifest.path(), deleteFiles);
-                }
-            } catch (Exception e) {
-                LOG.warn("Failed to read delete manifest {} for cache", manifest.path(), e);
-            }
-        }
-
-        // Return the original tasks (cache is already filled)
-        return tasks;
-    }
-
-    /**
-     * TaskContext helper class (replacement for ManifestGroup.TaskContext)
-     */
-    private static class TaskContextHelper {
-        private final String schemaAsString;
-        private final String specAsString;
-        private final ResidualEvaluator residuals;
-        private final boolean dropStats;
-        private final Set<Integer> columnsToKeepStats;
-        private final ScanMetrics scanMetrics;
-
-        TaskContextHelper(String schemaAsString, String specAsString,
-                ResidualEvaluator residuals,
-                boolean dropStats, Set<Integer> columnsToKeepStats,
-                ScanMetrics scanMetrics) {
-            this.schemaAsString = schemaAsString;
-            this.specAsString = specAsString;
-            this.residuals = residuals;
-            this.dropStats = dropStats;
-            this.columnsToKeepStats = columnsToKeepStats;
-            this.scanMetrics = scanMetrics;
-        }
-
-        String schemaAsString() {
-            return schemaAsString;
-        }
-
-        String specAsString() {
-            return specAsString;
-        }
-
-        ResidualEvaluator residuals() {
-            return residuals;
-        }
-
-        boolean shouldKeepStats() {
-            return !dropStats;
-        }
-
-        Set<Integer> columnsToKeepStats() {
-            return columnsToKeepStats;
-        }
-
-        ScanMetrics scanMetrics() {
-            return scanMetrics;
-        }
+        return tableName;
     }
 
     private Split createIcebergSplit(FileScanTask fileScanTask) {
